Make element's Comparable conformance optional
OSet can be much more widely applicable if it can be used with element types that don't conform to Comparable. This conformance is only needed in order to implement sort().

This change removes E's conformance to Comparable and adds an extension that implements sort() only when E conforms to Comparable.